Break up consensus into smaller files
