+++ .gitignore
@@ -14,5 +14,6 @@ Cargo.lock
 # MSVC Windows builds of rustc generate these, which store debugging information
 *.pdb
 
+.rocks/
 nodes/
 Procfile
