+++ src/consensus/mod.rs
@@ -1,2 +1,4 @@
 pub mod consensus;
+pub mod proposer;
 mod timers;
+pub mod validator;
